下单

+++ common/dm-common/dm-common-dao/src/main/java/cn/dm/mapper/DmOrderMapper.java
@@ -10,16 +10,16 @@ public interface DmOrderMapper {
 
 	public DmOrder getDmOrderById(@Param(value = "id") Long id)throws Exception;
 
-	public List<DmOrder>	getDmOrderListByMap(Map<String,Object> param)throws Exception;
+	public List<DmOrder>	getDmOrderListByMap(Map<String, Object> param)throws Exception;
 
-	public Integer getDmOrderCountByMap(Map<String,Object> param)throws Exception;
+	public Integer getDmOrderCountByMap(Map<String, Object> param)throws Exception;
 
-	public Integer insertDmOrder(DmOrder dmOrder)throws Exception;
+	public Long insertDmOrder(DmOrder dmOrder)throws Exception;
 
-	public Integer updateDmOrder(DmOrder dmOrder)throws Exception;
+	public Long updateDmOrder(DmOrder dmOrder)throws Exception;
 
 	public Integer deleteDmOrderById(@Param(value = "id") Long id)throws Exception;
 
-	public Integer batchDeleteDmOrder(Map<String,List<String>> params);
+	public Integer batchDeleteDmOrder(Map<String, List<String>> params);
 
 }
